Recibos funcionando Perfeitamente, da parcela e dos pagametnos

+++ View/FrmBaixarParcela.cs
@@ -207,7 +207,7 @@ namespace GVC.View
             }
             else
             {
-                var parcelasLong = _parcelasIds.Select(id => (long)id).ToList();
+                var parcelasLong = _parcelasIds.Select(id => (int)id).ToList();
 
                 parcelaBLL.BaixarParcelasEmLote(
                     parcelasLong,
